requirements.txt
boto3==1.4.4
botocore==1.4.92
dj-database-url==0.4.1
Django==1.10.4
djangorestframework==3.5.3
docutils==0.13.1
futures==3.0.5
gunicorn==19.6.0
jmespath==0.9.0
psycopg2==2.6.2
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
whitenoise==3.2.2
Pillow==3.4.0
requests==2.13.0